overlap non-sd gsm8k passing
