+++ python/sglang/srt/managers/scheduler.py
@@ -112,6 +112,7 @@ from sglang.srt.managers.io_struct import (
 from sglang.srt.managers.mm_utils import init_embedding_cache
 from sglang.srt.managers.schedule_batch import (
     FINISH_ABORT,
+    ModelWorkerBatch,
     MultimodalInputs,
     Req,
     ScheduleBatch,
@@ -134,10 +135,15 @@ from sglang.srt.mem_cache.hiradix_cache import HiRadixCache
 from sglang.srt.mem_cache.radix_cache import RadixCache
 from sglang.srt.mem_cache.swa_radix_cache import SWARadixCache
 from sglang.srt.metrics.collector import SchedulerMetricsCollector, SchedulerStats
-from sglang.srt.model_executor.forward_batch_info import ForwardMode, PPProxyTensors
+from sglang.srt.model_executor.forward_batch_info import (
+    ForwardBatchOutput,
+    ForwardMode,
+    PPProxyTensors,
+)
 from sglang.srt.reasoning_parser import ReasoningParser
 from sglang.srt.server_args import PortArgs, ServerArgs
-from sglang.srt.speculative.spec_info import SpecInfo, SpeculativeAlgorithm
+from sglang.srt.speculative.eagle_utils_v2 import EagleDraftInput
+from sglang.srt.speculative.spec_info import SpeculativeAlgorithm
 from sglang.srt.torch_memory_saver_adapter import TorchMemorySaverAdapter
 from sglang.srt.two_batch_overlap import TboDPAttentionPreparer
 from sglang.srt.utils import (
@@ -224,6 +230,119 @@ class IdleSleeper:
 
     def maybe_sleep(self):
         self.poller.poll(1000)
+
+
+class FutureMap:
+    def __init__(
+        self,
+        spec_algorithm: SpeculativeAlgorithm,
+        max_running_requests: int,
+        device: torch.device,
+    ):
+        self.spec_algorithm = spec_algorithm
+        self.future_ct = 0
+        self.future_limit = max_running_requests * 3
+        self.future_buffer_len = (
+            max_running_requests * 5
+        )  # 2 * max_running_requests for overlap schedule
+        self.device = device
+
+        if self.spec_algorithm.is_eagle():
+            self.future_topk_p_map = [None] * (self.future_buffer_len)
+            self.future_topk_index_map = [None] * (self.future_buffer_len)
+            self.future_hidden_states_map = [None] * (self.future_buffer_len)
+            self.future_verified_id_map = [None] * (self.future_buffer_len)
+            self.future_new_seq_lens_map = [None] * (self.future_buffer_len)
+        else:
+            self.future_next_token_ids_map = torch.empty(
+                (self.future_buffer_len,), dtype=torch.int64, device=self.device
+            )
+
+    def get_next_future_ct(self, bs: int) -> int:
+        cur_future_ct = self.future_ct
+        # increment the future_ct, circular buffer pointer
+        self.future_ct = (cur_future_ct + bs) % self.future_limit
+        return cur_future_ct
+
+    def resolve_future(self, model_worker_batch: ModelWorkerBatch):
+        if self.spec_algorithm.is_eagle():
+            spec_info = model_worker_batch.spec_info
+            if spec_info is None:
+                return
+
+            ids = [-idx for idx in spec_info.topk_p.tolist()]
+
+            topk_p_stacked = [self.future_topk_p_map[idx] for idx in ids]
+            topk_index_stacked = [self.future_topk_index_map[idx] for idx in ids]
+            hidden_states_stacked = [self.future_hidden_states_map[idx] for idx in ids]
+            verified_id_stacked = [self.future_verified_id_map[idx] for idx in ids]
+            new_seq_lens_stacked = [self.future_new_seq_lens_map[idx] for idx in ids]
+
+            # TODO: think if there's a good way to hide the stack overhead
+            # e.g. when the ids are continuous, we can just use a pointer to the first element
+            spec_info.topk_p = torch.stack(topk_p_stacked, dim=0)
+            spec_info.topk_index = torch.stack(topk_index_stacked, dim=0)
+            spec_info.hidden_states = torch.stack(hidden_states_stacked, dim=0)
+            spec_info.verified_id = torch.stack(verified_id_stacked, dim=0)
+            spec_info.new_seq_lens = torch.stack(new_seq_lens_stacked, dim=0)
+        else:
+            input_ids = model_worker_batch.input_ids
+            input_ids[:] = torch.where(
+                input_ids < 0,
+                self.future_next_token_ids_map[torch.clamp(-input_ids, min=0)],
+                input_ids,
+            )
+
+    def update_next_future(
+        self,
+        schedule_batch: ScheduleBatch,
+        future_ct: int,
+        bs: int,
+        allocate_lens: torch.Tensor,
+    ):
+        if self.spec_algorithm.is_eagle():
+            # future_spec_info_indices is a reference to the next spec_info fields, val is a reference id stored as a negative index.
+            future_spec_info_indices = torch.arange(
+                -(future_ct + 1),
+                -(future_ct + 1 + bs),
+                -1,
+                dtype=torch.int64,
+                device=self.device,
+            )
+
+            schedule_batch.spec_info = EagleDraftInput(
+                topk_p=future_spec_info_indices,
+                topk_index=future_spec_info_indices,
+                hidden_states=future_spec_info_indices,
+                verified_id=future_spec_info_indices,
+                new_seq_lens=future_spec_info_indices,
+                allocate_lens=allocate_lens,  # allocate_lens is never a future
+            )
+        else:
+            return torch.arange(
+                -(future_ct + 1),
+                -(future_ct + 1 + bs),
+                -1,
+                dtype=torch.int64,
+                device=self.device,
+            )
+
+    def store_to_map(self, future_ct: int, bs: int, forward_output: ForwardBatchOutput):
+        if self.spec_algorithm.is_eagle():
+            spec_info = forward_output.spec_info
+            # Store references (no clone) for each request into circular buffers
+            # NOTE: self.future_allocate_lens_map is not needed here because it's assigned to batch.spec_info.allocate_lens, not a future
+            for i in range(bs):
+                slot = future_ct + i + 1
+                self.future_topk_p_map[slot] = spec_info.topk_p[i]
+                self.future_topk_index_map[slot] = spec_info.topk_index[i]
+                self.future_hidden_states_map[slot] = spec_info.hidden_states[i]
+                self.future_verified_id_map[slot] = spec_info.verified_id[i]
+                self.future_new_seq_lens_map[slot] = spec_info.new_seq_lens[i]
+        else:
+            self.future_next_token_ids_map[future_ct + 1 : future_ct + bs + 1] = (
+                forward_output.next_token_ids
+            )
 
 
 class Scheduler(
@@ -495,8 +614,6 @@ class Scheduler(
 
         # init future map for overlap schedule
         if self.enable_overlap:
-            from sglang.srt.speculative.eagle_utils_v2 import FutureMap
-
             self.future_map = FutureMap(
                 spec_algorithm=self.spec_algorithm,
                 max_running_requests=self.max_running_requests,
@@ -1524,16 +1641,16 @@ class Scheduler(
             protected_size = self.tree_cache.protected_size()
 
             total_kv_indices = set(range(1, self.max_total_num_tokens + 1))
-            evictable_kv_indices = set(self.tree_cache.evictable_kv_indices())
+            cached_kv_indices = set(self.tree_cache.cached_kv_indices())
             available_kv_indices = set(
                 self.token_to_kv_pool_allocator.free_pages.tolist()
             )
 
             kv_indices_leak = total_kv_indices != (
-                evictable_kv_indices | available_kv_indices
+                cached_kv_indices | available_kv_indices
             )
             missing_kv_indices = total_kv_indices - (
-                evictable_kv_indices | available_kv_indices
+                cached_kv_indices | available_kv_indices
             )
 
             memory_leak = (available_size + evictable_size) != (
@@ -1543,7 +1660,7 @@ class Scheduler(
             ) or kv_indices_leak
             token_msg = f"{self.max_total_num_tokens=}, {available_size=}, {evictable_size=}, {protected_size=}, {kv_indices_leak=}\n"
             if kv_indices_leak:
-                token_msg += f"{len(total_kv_indices)=}, {len(evictable_kv_indices)=}, {len(available_kv_indices)=}, {len(missing_kv_indices)=}, {total_kv_indices=}, {evictable_kv_indices=}, {available_kv_indices=}, {missing_kv_indices=}\n"
+                token_msg += f"{len(total_kv_indices)=}, {len(cached_kv_indices)=}, {len(available_kv_indices)=}, {len(missing_kv_indices)=}, {total_kv_indices=}, {cached_kv_indices=}, {available_kv_indices=}, {missing_kv_indices=}\n"
 
         if memory_leak:
             msg = "token_to_kv_pool_allocator memory leak detected! " f"{token_msg}"
@@ -1916,16 +2033,18 @@ class Scheduler(
 
                 # Run forward in a separate stream to avoid blocking the main stream.
                 with self.forward_stream_ctx:
-                    # model_worker_batch.spec_info is a future from the previous batch
+                    # spec_info / input_ids is a future from the previous batch.
                     # resolve it with the future_map
                     self.future_map.resolve_future(model_worker_batch)
 
                     forward_output = self.forward_worker.forward_batch_generation(
                         model_worker_batch
                     )
-                    # store the spec_info to the future_map for next batch resolution
-                    if forward_output.spec_info is not None:
-                        self.future_map.store_to_map(future_ct, bs, forward_output)
+
+                    # store spec_info / output_ids to the future_map for next batch resolution
+                    self.future_map.store_to_map(future_ct, bs, forward_output)
+
+                    # copy the output to the main stream
                     copy_done = torch.cuda.Event()
                     copy_done.record()
 
@@ -1935,6 +2054,8 @@ class Scheduler(
                     if forward_output.spec_info is not None
                     else None
                 )
+
+                # update the spec_info / output_ids for next batch resolution
                 self.future_map.update_next_future(
                     batch, future_ct, bs, output_allocate_lens
                 )
