+++ python/sglang/srt/utils.py
@@ -2893,11 +2893,6 @@ def parse_module_path(module_path, function_name, create_dummy):
 
     return final_module, None
 
-def alloc_len_per_eagle_decode(worker: "EagleWorker") -> int:
-    return max(
-        worker.num_steps * worker.topk,
-        worker.num_draft_tokens,
-    )
 
 # LoRA-related constants and utilities
 SUPPORTED_LORA_TARGET_MODULES = [
